Rename, Remove/inquiryRegistration: 중복 MemberRepository 삭제, MemberQuestion 폴더 구조 변경
